Update run-server.sh
Only copy application if it does not exist

+++ run-server.sh
@@ -1,7 +1,9 @@
 #!/bin/sh
 
-cd /opt/foundryvtt
-cp /host/foundryvtt*.zip .
-unzip -o foundryvtt*.zip
-rm foundryvtt*.zip
+if [! -d "/opt/foundryvtt" ]
+  cd /opt/foundryvtt
+  cp /host/foundryvtt*.zip .
+  unzip -o foundryvtt*.zip
+  rm foundryvtt*.zip
+fi
 node resources/app/main.js --dataPath=/data/foundryvtt
